Create separate `ImageStats` for images
