+++ crates/viewer/re_viewer_context/src/tensor/mod.rs
@@ -1,11 +1,13 @@
 // TODO(andreas): Move tensor utilities to a tensor specific crate?
 
 mod image_decode_cache;
+mod image_stats;
 mod image_stats_cache;
 mod tensor_stats;
 mod tensor_stats_cache;
 
 pub use image_decode_cache::ImageDecodeCache;
+pub use image_stats::ImageStats;
 pub use image_stats_cache::ImageStatsCache;
 pub use tensor_stats::TensorStats;
 pub use tensor_stats_cache::TensorStatsCache;
